+++ docs/.vitepress/config.ts
@@ -28,7 +28,7 @@ export default {
         activeMatch: "/showcases/",
       },
      {
-        text: "2.9.8.7",
+        text: "2.9.8.8",
         items: [
           {
             text: 'Changelog',
@@ -90,7 +90,8 @@ export default {
           // collapsed: false,
           items: [
             { text: "Overview", link: "/getting-started/overview" },
-            { text: "Install", link: "/getting-started/install" },
+            { text: "Installation", link: "/getting-started/install" },
+            { text: "Migration", link: "/getting-started/migrate" },
             { text: "Dependencies", link: "/getting-started/dependencies" },
             { text: "Update", link: "/getting-started/update" },
           ],
@@ -123,6 +124,7 @@ export default {
           items: [
             { text: "Dotfiles Customization", link: "/customization/dotfiles" },
             { text: "Config Variants", link: "/customization/variants" },
+            { text: "Display Manager", link: "/customization/displaymanager" },
             { text: "Customize Waybar", link: "/customization/waybar" },
             { text: "Shell (Zsh & Bash)", link: "/customization/shell" },
             { text: "Default Terminal", link: "/customization/terminal" },